Separating core admin behaviour from theme-specific behaviour
Changes to better support admin themes that want to do something
different from the default behaviours for tab reordering, extend drag
and drop, and field expanding.

Behaviours of the default theme considered "not core" have been moved
into a new file that resides in the theme's folder and runs immediately
after admin.js.php.
