+++ admin/themes/default/layout.twig
@@ -12,6 +12,7 @@
         <script src="$site.chyrp_url/includes/lib/gz.php?file=plugins.js" type="text/javascript" charset="utf-8"></script>
         <script src="$site.chyrp_url/includes/lib/gz.php?file=redactor/redactor.min.js" type="text/javascript" charset="utf-8"></script>
         <script src="$site.chyrp_url/includes/admin.js.php?action=$route.action" type="text/javascript" charset="utf-8"></script>
+        <script src="$theme_url/javascript.php?action=$route.action" type="text/javascript" charset="utf-8"></script>
 ${ trigger.call("admin_head") }
     </head>
     <body>
